+++ demo/src/main/java/com/icici/demo/repos/AccountsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.icici.demo.entities.Accounts;
 
-public interface AccountsRepository extends JpaRepository<Accounts, Integer> {
+public interface AccountsRepository extends JpaRepository<Accounts, Long> {
     
 }
